workspace/finalytics_spark/pyproject.toml
[tool.poetry]
name = "finalytics-spark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
pyyaml = "^6.0.2"
yfinance = "^0.2.51"
psycopg2-binary = "^2.9.10"
nbimporter = "^0.3.4"
nbformat = "^5.10.4"
pyspark = "3.5.2"
yahooquery = "^2.3.7"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
